Move ChannelThumbnailView into views.py and delete content/api.py
The logger viewsets imported OptionalPageNumberPagination from content/api.py,
which held a copy of the kolibri.core.utils.pagination original; they now take
it from there.
