+++ docs/backend_architecture/content/api_methods.rst
@@ -1,5 +1,40 @@
 API methods
 -----------
 
-.. automodule:: kolibri.core.content.api
+.. One entry per content module routed from content/api_urls.py or public/api_urls.py.
+
+.. automodule:: kolibri.core.content.viewsets.channel_metadata
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.content_request
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.contentnode.base
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.contentnode.bookmarks
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.contentnode.granular
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.contentnode.progress
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.contentnode.tree
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.contentnode.user
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.import_metadata
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.remote_channel
+    :members:
+
+.. automodule:: kolibri.core.content.viewsets.share_file
+    :members:
+
+.. automodule:: kolibri.core.content.views
     :members:
